Dialouge
Added the class TextLoader to easily implement new dialouge. The easiest
use is by a static method but it is also possible to make a textloader
class and change the file by naming a new file. Next I need to improve
the conversation code because it uses a class-global variable. I need
more and different variables